+++ grpc.go
@@ -1,0 +1,139 @@
+package kvs
+
+import (
+	"context"
+	"errors"
+	"fmt"
+	"time"
+
+	"github.com/Jille/raft-grpc-leader-rpc/rafterrors"
+	pb "github.com/bootjp/kvs/proto"
+)
+
+func (r RPCInterface) DeleteData(_ context.Context, req *pb.DeleteRequest) (*pb.DeleteResponse, error) {
+	if len(req.GetKey()) > KeyLimit {
+		return &pb.DeleteResponse{
+			Status:      pb.Status_ABORT,
+			CommitIndex: r.Raft.AppliedIndex(),
+		}, fmt.Errorf("reachd key size limit: %d max key size %d", len(req.Key), KeyLimit)
+	}
+
+	var tmp [KeyLimit]byte
+	copy(tmp[:], req.Key)
+
+	pair := Pair{Key: &tmp, Value: nil, IsDelete: true}
+	e, err := EncodePair(pair)
+	if err != nil {
+		return &pb.DeleteResponse{
+			Status: pb.Status_ABORT,
+		}, err
+	}
+
+	f := r.Raft.Apply(e, time.Second)
+	if err := f.Error(); err != nil {
+		return &pb.DeleteResponse{
+			CommitIndex: f.Index(),
+			Status:      pb.Status_ABORT,
+		}, errors.Unwrap(rafterrors.MarkRetriable(err))
+	}
+
+	return &pb.DeleteResponse{
+		CommitIndex: f.Index(),
+		Status:      pb.Status_COMMIT,
+	}, nil
+}
+
+func (r RPCInterface) AddData(_ context.Context, req *pb.AddDataRequest) (*pb.AddDataResponse, error) {
+	if len(req.GetKey()) > KeyLimit {
+		return &pb.AddDataResponse{
+			Status:      pb.Status_ABORT,
+			CommitIndex: r.Raft.AppliedIndex(),
+		}, fmt.Errorf("reachd key size limit: %d max key size %d", len(req.Key), KeyLimit)
+	}
+
+	var tmp [KeyLimit]byte
+	copy(tmp[:], req.Key)
+
+	pair := Pair{Key: &tmp, Value: &req.Data, Expire: TTLtoTime(req.GetTtl().AsDuration())}
+	e, err := EncodePair(pair)
+	if err != nil {
+		return &pb.AddDataResponse{
+			Status: pb.Status_ABORT,
+		}, err
+	}
+
+	f := r.Raft.Apply(e, time.Second)
+	if err := f.Error(); err != nil {
+		return &pb.AddDataResponse{
+			CommitIndex: f.Index(),
+			Status:      pb.Status_ABORT,
+		}, errors.Unwrap(rafterrors.MarkRetriable(err))
+	}
+
+	resp := f.Response()
+	if err, ok := resp.(error); ok {
+		return &pb.AddDataResponse{
+			CommitIndex: f.Index(),
+			Status:      pb.Status_ABORT,
+		}, errors.Unwrap(rafterrors.MarkRetriable(err))
+	}
+
+	ff := r.Raft.Barrier(1 * time.Second)
+	if err := ff.Error(); err != nil {
+		return &pb.AddDataResponse{
+			CommitIndex: f.Index(),
+			Status:      pb.Status_ABORT,
+		}, errors.Unwrap(rafterrors.MarkRetriable(err))
+	}
+
+	return &pb.AddDataResponse{
+		CommitIndex: f.Index(),
+		Status:      pb.Status_COMMIT,
+	}, nil
+}
+
+func (r RPCInterface) GetData(_ context.Context, req *pb.GetDataRequest) (*pb.GetDataResponse, error) {
+	r.KVS.mtx.RLock()
+	defer r.KVS.mtx.RUnlock()
+
+	if len(req.GetKey()) > KeyLimit {
+		return &pb.GetDataResponse{
+			Key:         req.Key,
+			Data:        nil,
+			Error:       pb.GetDataError_FETCH_ERROR,
+			ReadAtIndex: r.Raft.AppliedIndex(),
+		}, fmt.Errorf("reachd key size limit: %d max key size %d", len(req.Key), KeyLimit)
+	}
+
+	var tmp [KeyLimit]byte
+	copy(tmp[:], req.Key)
+
+	v, ok := r.KVS.data[tmp]
+	if !ok {
+		return &pb.GetDataResponse{
+			Key:         req.Key,
+			Data:        nil,
+			Error:       pb.GetDataError_DATA_NOT_FOUND,
+			ReadAtIndex: r.Raft.AppliedIndex(),
+		}, nil
+	}
+
+	// check expire
+	if !v.Expire.NoExpire && v.Expire.Expire(time.Now().UTC()) {
+		r.gcc <- *v
+		return &pb.GetDataResponse{
+			Key:         req.Key,
+			Data:        nil,
+			Error:       pb.GetDataError_DATA_NOT_FOUND,
+			ReadAtIndex: r.Raft.AppliedIndex(),
+		}, nil
+	}
+
+	return &pb.GetDataResponse{
+		Key:         req.Key,
+		Data:        *v.Value,
+		Error:       pb.GetDataError_NO_ERROR,
+		ReadAtIndex: r.Raft.AppliedIndex(),
+	}, nil
+
+}
